Rename skse.h to jc_skse.h because preprocessor ignores this file completely
