+++ src/JContainers/src/skse/papyrus_args.hpp
@@ -2,7 +2,7 @@
 
 #include <common/ITypes.h>
 #include <SKSE/SKSE.h>
-#include "skse/skse.h"
+#include "skse/jc_skse.h"
 #include "reflection/tes_binding.h"
 
 // Different meachanism used in CommonLibNG-SE to identify the types, these methods are now obsolete.
